requirements.txt
# pip install -r requirements.txt

numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
PyYAML>=6.0